fix: added evaluateDeep into evaluate
